requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1

# Browser automation — local (SOS) + cloud (company/person)
browser-use>=0.12.6
browser-use-sdk>=2.0.0
google-genai>=1.0.0

# Document parsing
pdfplumber==0.11.0
openpyxl==3.1.2
